common, window: stricter audio mimetype validation
Technically, taglib supports a lot of filetypes, but there may be
implementation details that are different between files.

To avoid the additional maintainership effort of maintaining exotic
formats, this commit makes the mimetype validation stricter, only
allowing some of the most common filetypes.

It also updates the .desktop file to remove a few additional formats
that we don't support.