identify: redesigned MusicBrainz identification dialog
The original identification dialog used AcoustID to identify songs. However, the approach was less than ideal - we always selected the first best matching release, and many songs weren't in AcoustID's database in the first place. As such, the feature worked rather poorly.

This commit changes this in multiple ways:
- The primary method of getting information about tracks is now through a MusicBrainz query, based on the metadata of the file. AcoustID is still used as a fallback.
- A set of algorithms has been added to attempt to guess the correct release to use for each track, and to group together similar releases to hopefully reduce most misidentifications.

This is a massive change, and will require lots of testing. Suggestions and improvements are welcome!