window: port to new libadwaita layout widgets
Move sidebar code into the window, since we need to have the search pane
and bottom select all/remove selected menubar in the main window UI for it
to be part of the AdwToolbarView. As with many last-minute changes, this is
probably going to cause some dumb bugs so testing is appreciated.