musicbrainz: implement async loading for covers/thumbnails
Allow thumbnails to be loaded independently of remaining MusicBrainz
requests. This somewhat reduces the waiting time for thumbnails, thus
adding a nice speedup to large identify queries, and makes thumbnails
for alternative releases only load once the popup is opened, which
can massively speed up load times for releases with many alternatives.

A potential improvement would be to create a queue for cover downloads,
especially in the apply step (right now we still treat the operation
as blocking), and allow 3 or so covers to be downloaded at once. Right
now, all thumbnails/covers download with no timeout all at once, which
might saturate slower connections.

Part of #131.