tagentry, fileview: rework file binding/unbinding
Another huge refactor, yay! This one makes selecting files *way* faster -
we're talking ~3 seconds to select ~9800 files, instead of a few minutes(!).

Out of those 3 seconds, only ~2s is the binding part (the remaining second
is mostly taken up by repeated is_selected calls from the sidebar filelist
items, which I'll try to reduce in another commit).

Unbinding for the same amount of files takes ~0.3s (0.2s is spent in the
unbind function).

The last refactor introduced a new class for entries bound to tags: the
EartagTagEntryBase class and its subclasses. The idea was that each entry
managed its own property. It also had a very simple API - you called
bind_to_files to bind to a list of files, and unbind_from_files to unbind.
That same API could be reused for other entries that need to count file
counts, but aren't tag entries - see the AlbumCoverButton, FilenameEntry
and FileInfo.

However, it turns out that this was very inefficient. Turns out that
making the bindings multiple times was slow, and we were probably doing
more loops and general work than we needed to.

Thus, the entries no longer manage the connection to the file, or even
anything related to setting values! All of this work has been moved to
the EartagTagEntryManager.

The manager binds to a list of entries and listens to input/applies values
from files, then binds to files and listens to their changes/applies the
values.

For each file, there is only *one* connection - to the "modified" signal,
which is then routed to the appropriate entry based on the modified property.

Another surprising major optimization was switching from lists to sets for
file storage in all classes that bound to files. It turns out that sets are
much faster for lookups, adding items, and even removing items - not to mention
that the set object API allows us to do things like find set differences with
a simple "-" operand. That optimization alone cut down perhaps the most
significant slowdowns - I saw one of the non-entry classes' unbind_from_file
function go from taking ~9 seconds to... far less, probably sub-1s.